feat(ui): header rework
