+++ src/feature/About/styles.module.scss
@@ -1,6 +1,5 @@
 .root {
   min-height: 100vh;
-  padding: 80px 0px;
 }
 
 .content {
