+++ src/feature/Header/NavMobile/styles.module.scss
@@ -1,0 +1,94 @@
+.menuToggle {
+  display: block;
+  position: fixed;
+  top: 15px;
+  right: 30px;
+  z-index: 3;
+
+  .line {
+    position: relative;
+    width: 33px;
+    height: 4px;
+    z-index: 2;
+    margin-bottom: 5px;
+    background: var(--primary-text-color);
+    border-radius: 3px;
+    transition: transform 0.5s cubic-bezier(0.77, 0.2, 0.05, 1),
+      background-color 0.5s cubic-bezier(0.77, 0.2, 0.05, 1), opacity 0.5s ease;
+    transform-origin: 4px 0px;
+
+    &:first-child {
+      transform-origin: 0% 0%;
+    }
+    &:nth-last-child(2) {
+      transform-origin: 0% 100%;
+    }
+  }
+
+  @media (hover: hover) {
+    &:hover {
+      .line {
+        background: var(--secondary-text-color);
+      }
+    }
+  }
+}
+
+.menuCheckbox {
+  display: block;
+  width: 40px;
+  height: 32px;
+  position: absolute;
+  top: -7px;
+  left: -5px;
+  cursor: pointer;
+  opacity: 0;
+  z-index: 4;
+}
+
+.menu {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+
+  padding: 100px 50px;
+  box-sizing: border-box;
+
+  background: #ededed;
+  list-style: none;
+  transform: translateX(100%);
+  transition: transform 0.5s cubic-bezier(0.77, 0.2, 0.05, 1);
+
+  li {
+    display: flex;
+    justify-content: end;
+    padding: 20px 0;
+    font-size: 22px;
+    font-size: 48px;
+
+    label {
+      cursor: pointer;
+      color: #232323;
+    }
+  }
+}
+
+.menuCheckbox:checked ~ .line {
+  opacity: 1;
+  transform: rotate(45deg) translate(-2px, -1px);
+  background: #232323;
+
+  &:nth-last-child(2) {
+    transform: rotate(-45deg) translate(0, -1px);
+  }
+  &:nth-last-child(3) {
+    opacity: 0;
+    transform: rotate(0deg) scale(0.2, 0.2);
+  }
+}
+
+.menuCheckbox:checked ~ .menu {
+  transform: none;
+}
