+++ src/shared/ScrollDownBlock/styles.module.scss
@@ -33,7 +33,7 @@
 
   &.visible {
     opacity: 1;
-    transform: translateY(0px);
+    transform: translateY(-60px);
   }
 }
