+++ src/shared/SectionHeader/styles.module.scss
@@ -2,4 +2,5 @@
   display: flex;
   justify-content: center;
   font-size: var(--title-font-size);
+  padding: 105px 0 50px 0;
 }
