+++ src/styles/global.scss
@@ -4,6 +4,7 @@
 
   --body-bg-color: #1e1e1e;
   --components-bg-color: #2f2f45;
+  --components-hover-color: #41415a;
 
   --cta-color: #0071e3;
   --cta-hover-color: #0077ed;
